pull ProviderLoader from VerbBuilder
